+++ Library/FSGIMG/FSGIMGPackage.cs
@@ -140,7 +140,7 @@ namespace GameArchives.FSGIMG
         }
         else if (filename[0] == 'F')
         {
-          ret.AddFile(new FSGIMGFile(real_name, ret, filestream, (ulong)desc.data_offset, (ulong)desc.size));
+          ret.AddFile(new FSGIMGFile(real_name, ret, filestream, desc.data_offset, desc.size));
         }
         else
         {
